Fixed LED blinking.
Noticed earlier when running this script that it wasn't blinking the LED properly on my turtle and I was getting warnings about the earlier LED path not being valid. So instead I looked around in a similar path when i found that `/sys/class/leds/lan-turtle\:orange\:system/brightness` was valid.