Corrected Responder version and adjusted unpacking
This changes the version to 3.0.6.0, which works with Python 2. Huge shout out to dark_pyrro for sharing their knowledge and catching this. I've also adjusted the unpacking and handling based on their suggestion, in addition to removing the .exe deletion, as it no longer ships with the .exe. 

To validate functionality, I deleted the Responder folder on the SD card, the quickcreds loot folder, and ensured the /etc/turtle/modules/quickcreds file was up to date with this commit and confirmed that I received creds.